+++ tests/testthat/test-recode.R
@@ -14,13 +14,11 @@ test_that("Recode discrete values into groups with all values covered", {
         "55 to under 65" = 55:64,
         "65 and older"   = 65:100))
 
-    recode_df <- suppressMessages(test_df |> recode("age_group", age = age.))
+    test_df[["age_group"]] <- suppressMessages(test_df |> recode(age = age.))
 
-    expect_true(ncol(recode_df) == ncol(test_df) + 1)
+    expect_true("age_group" %in% names(test_df))
 
-    expect_true("age_group" %in% names(recode_df))
-
-    expect_true(length(unique(recode_df[["age_group"]])) <= 6)
+    expect_true(length(unique(test_df[["age_group"]])) <= 6)
 })
 
 
@@ -31,14 +29,12 @@ test_that("Recode interval values into groups with all values covered", {
         "1000 to under 2000" = 1000:1999,
         "2000 and more"      = 2000:99999))
 
-    test_df   <- test_df |> collapse::fsubset(!is.na(income))
-    recode_df <- suppressMessages(test_df |> recode("income_group", income = income.))
+    test_df <- test_df |> collapse::fsubset(!is.na(income))
+    test_df[["income_group"]] <- suppressMessages(test_df |> recode(income = income.))
 
-    expect_true(ncol(recode_df) == ncol(test_df) + 1)
+    expect_true("income_group" %in% names(test_df))
 
-    expect_true("income_group" %in% names(recode_df))
-
-    expect_true(length(unique(recode_df[["income_group"]])) <= 5)
+    expect_true(length(unique(test_df[["income_group"]])) <= 5)
 })
 
 
@@ -49,9 +45,7 @@ test_that("Recode interval values with NA values", {
         "1000 to under 2000" = 1000:1999,
         "2000 and more"      = 2000:99999))
 
-    expect_message(recode_df <- test_df |> recode("income_group", income = income.), " ! WARNING: Variable 'income' has NA values")
-
-    expect_true("income_group" %in% names(recode_df))
+    expect_message(test_df[["income_group"]] <- test_df |> recode(income = income.), " X ERROR: Variable 'income' has NA values")
 })
 
 
@@ -62,15 +56,13 @@ test_that("Recode discrete values into groups with not all values covered", {
         "55 to under 65" = 55:64,
         "65 and older"   = 65:100))
 
-    test_df   <- test_df |> collapse::fsubset(!is.na(income))
-    recode_df <- suppressMessages(test_df |> recode("age_group", age = age.))
+    test_df <- test_df |> collapse::fsubset(!is.na(income))
+    test_df[["age_group"]] <- suppressMessages(test_df |> recode(age = age.))
 
-    expect_true(ncol(recode_df) == ncol(test_df) + 1)
+    expect_true("age_group" %in% names(test_df))
 
-    expect_true("age_group" %in% names(recode_df))
-
-    unique_age       <- length(unique(recode_df[["age"]]))
-    unique_age_group <- length(unique(recode_df[["age_group"]]))
+    unique_age       <- length(unique(test_df[["age"]]))
+    unique_age_group <- length(unique(test_df[["age_group"]]))
     expect_true(unique_age_group >= 4 & unique_age_group < unique_age)
 })
 
@@ -81,15 +73,13 @@ test_that("Recode interval values into groups with not all values covered", {
         "500 to under 1000"  = 500:999,
         "1000 to under 2000" = 1000:1999))
 
-    test_df   <- test_df |> collapse::fsubset(!is.na(income))
-    recode_df <- suppressMessages(test_df |> recode("income_group", income = income.))
+    test_df <- test_df |> collapse::fsubset(!is.na(income))
+    test_df[["income_group"]] <- suppressMessages(test_df |> recode(income = income.))
 
-    expect_true(ncol(recode_df) == ncol(test_df) + 1)
+    expect_true("income_group" %in% names(test_df))
 
-    expect_true("income_group" %in% names(recode_df))
-
-    unique_income       <- length(unique(recode_df[["income"]]))
-    unique_income_group <- length(unique(recode_df[["income_group"]]))
+    unique_income       <- length(unique(test_df[["income"]]))
+    unique_income_group <- length(unique(test_df[["income_group"]]))
     expect_true(unique_income_group >= 4 & unique_income_group < unique_income)
 })
 
@@ -98,18 +88,15 @@ test_that("Providing none format data frames", {
     age1 <- c(1, 2, 3, 4)
     age2 <- c("a", "b", "c", "d")
 
-    expect_message(recode_df1 <- test_df |> recode("age_group", age = age1), " X ERROR: The format for 'age' must be a data table")
-    expect_message(recode_df2 <- test_df |> recode("age_group", age = age2), " X ERROR: The format for 'age' must be a data table")
+    expect_message(test_df[["age_group"]] <- test_df |> recode(age = age1), " X ERROR: The format for 'age' must be a data table")
+    expect_message(test_df[["age_group"]] <- test_df |> recode(age = age2), " X ERROR: The format for 'age' must be a data table")
 
-    expect_identical(recode_df1, test_df)
-    expect_identical(recode_df2, test_df)
-
-    expect_true(!"age_group" %in% names(recode_df1))
-    expect_true(!"age_group" %in% names(recode_df2))
+    expect_true(!"age_group" %in% names(test_df))
+    expect_true(!"age_group" %in% names(test_df))
 })
 
 
-test_that("Given variable is no in data frame", {
+test_that("Given variable is not in data frame", {
     age. <- suppressMessages(discrete_format(
         "under 18"       = 0:17,
         "18 to under 25" = 18:24,
@@ -117,27 +104,9 @@ test_that("Given variable is no in data frame", {
         "55 to under 65" = 55:64,
         "65 and older"   = 65:100))
 
-    expect_message(recode_df <- test_df |> recode("age_group", dog = age.), " X ERROR: Variable 'dog' not found in the input data frame")
+    expect_message(test_df[["age_group"]] <- test_df |> recode(dog = age.), " X ERROR: Variable 'dog' not found in the input data frame")
 
-    expect_identical(recode_df, test_df)
-
-    expect_true(!"age_group" %in% names(recode_df))
-})
-
-
-test_that("Entering none character as new variable name", {
-    age. <- suppressMessages(discrete_format(
-        "under 18"       = 0:17,
-        "18 to under 25" = 18:24,
-        "25 to under 55" = 25:54,
-        "55 to under 65" = 55:64,
-        "65 and older"   = 65:100))
-
-    recode_df <- suppressMessages(test_df |> recode(123, age = age.))
-
-    expect_true(ncol(recode_df) == ncol(test_df) + 1)
-
-    expect_true("123" %in% names(recode_df))
+    expect_true(!"age_group" %in% names(test_df))
 })
 
 
@@ -150,11 +119,11 @@ test_that("Recoding with multilabel gives a warning", {
         "55 to under 65" = 55:64,
         "65 and older"   = 65:100))
 
-    expect_message(recode_df <- test_df |> recode("age_group", age = age.), " ! WARNING: The format for 'age' is a multilabel")
+    expect_message(test_df[["age_group"]] <- test_df |> recode(age = age.), " ! WARNING: The format for 'age' is a multilabel")
 })
 
 
-test_that("Recode won't overwrite existing variable", {
+test_that("Recode will overwrite existing variable", {
     age. <- suppressMessages(discrete_format(
         "under 18"       = 0:17,
         "18 to under 25" = 18:24,
@@ -162,7 +131,13 @@ test_that("Recode won't overwrite existing variable", {
         "55 to under 65" = 55:64,
         "65 and older"   = 65:100))
 
-    expect_message(recode_df <- test_df |> recode("age", age = age.), " X ERROR: Variable 'age' already exists")
+    test_df[["age"]] <- test_df |> recode(age = age.)
+
+    expect_true(all(c("under 18",
+                      "18 to under 25",
+                      "25 to under 55",
+                      "55 to under 65",
+                      "65 and older") %in% test_df[["age"]]))
 })
 
 ###############################################################################
@@ -180,7 +155,7 @@ test_that("Recode a variable with a multilabel format (discrete)", {
 
     recode_df <- test_df |> recode_multi(age = age.)
 
-    expect_true(ncol(recode_df) == ncol(test_df))
+    expect_true(collapse::fncol(recode_df) == collapse::fncol(test_df))
     expect_true(nrow(recode_df) > nrow(test_df))
 })
 
@@ -195,7 +170,7 @@ test_that("Recode a variable with a multilabel format (interval)", {
 
     recode_df <- test_df |> recode_multi(income = income.)
 
-    expect_true(ncol(recode_df) == ncol(test_df))
+    expect_true(collapse::fncol(recode_df) == collapse::fncol(test_df))
     expect_true(nrow(recode_df) > nrow(test_df))
 })
 
@@ -218,6 +193,6 @@ test_that("Recode multiple variables at once", {
 
     recode_df <- test_df |> recode_multi(age = age., income = income.)
 
-    expect_true(ncol(recode_df) == ncol(test_df))
+    expect_true(collapse::fncol(recode_df) == collapse::fncol(test_df))
     expect_true(nrow(recode_df) > nrow(test_df))
 })